New benchmark pages and matvec bench
